django_audio_waveform/requirements.txt
# Django Audio Waveform Requirements
Django>=4.2.0
djangorestframework>=3.14.0
Pillow>=10.0.0
numpy>=1.24.0
supabase>=2.0.0
cloudinary>=1.36.0
librosa>=0.10.0  # Optional: for real audio analysis
